+++ src/onegov.event/HISTORY.rst
@@ -7,6 +7,9 @@ Changelog
 - Adds ical export to occurrence collection.
   [msom]
 
+- Adds ical import to event collection.
+  [msom]
+
 - Adds tags, timestamp, uid and coordinates to ical exports.
   [msom]